New method for reading Sentinel-2 metadata
S2 metadata is now read in a dedicated function in the script:
read_satellite_metadata.py
It fixes the issue that a S2 scene is divided in to granules. Now we can
access the metadata of each granule and ensure that the right meatadata
is used in the atmospheric correction.
In the future is should be considered to read the metadata from other
sensors in this way also. This will limit code repetitions.